老师查看分数

+++ upload/models.py
@@ -22,4 +22,4 @@ class Record(models.Model):
         (Downloaded, '已下载'),
     )
     status = models.IntegerField(choices=Status_Choice, default=ToJudge)
-    Scores = models.IntegerField(default=0)
+    Scores = models.FloatField(default=0)
